requirements.txt
assemblyai
streamlit
google-cloud-aiplatform
google-cloud-storage
google-cloud-speech
python-dotenv
pydub